customers-database/requirements.txt
grpcio==1.78.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.46
typing-extensions==4.15.0
grpcio-tools==1.78.0
protobuf==6.33.5
setuptools==82.0.0
pytest